+++ CHANGELOG.md
@@ -66,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Build version in ``sat showrev`` is now read from the /etc/cray-release file
   and the field now reads "Release version".
+- Slurm version now checked via pod.
 
 ### Removed
 - Removed ``--substr`` option from ``sat showrev``, the effect of which can